add "save" button to default links and fix drag and drop reorder
I noticed the reorder wasn't working in mobile on the sidebar links editor
I fixed that for NewFileTemplate reorder, so, now I made a util to avoid do the same thing
on both files
