+++ frontend/src/components/prompts/NewFileTemplate.vue
@@ -4,18 +4,18 @@
     <div
       v-for="(item, index) in localItems"
       :key="index"
-      :ref="el => itemRefs[index] = el"
+      :ref="el => dragReorder.itemRefs[index] = el"
       class="new-file-template-row"
-      :class="{ 'is-dragging': draggingIndex === index }"
+      :class="{ 'is-dragging': dragReorder.draggingIndex === index }"
     >
       <i
         class="material-symbols drag-handle"
         aria-hidden="true"
-        @pointerdown="onPointerDown(index, $event)"
-        @pointermove="onPointerMove"
-        @pointerup="onPointerUp"
-        @pointercancel="onPointerCancel"
-        @lostpointercapture="onPointerCancel"
+        @pointerdown="dragReorder.onPointerDown(index, $event)"
+        @pointermove="dragReorder.onPointerMove($event)"
+        @pointerup="dragReorder.onPointerUp($event)"
+        @pointercancel="dragReorder.onPointerCancel($event)"
+        @lostpointercapture="dragReorder.onPointerCancel($event)"
       >drag_indicator</i>
       <i
         class="file-type-icon"
@@ -90,8 +90,8 @@
 
 <script>
 import { getters, mutations } from "@/store";
-import { getObjectProperty } from '@/utils/object.js';
 import { getTypeInfoFromExt } from "@/utils/mimetype";
+import { createDragReorder } from "@/utils/dragAndDropReorder.js";
 
 export default {
   name: "new-file-template",
@@ -105,24 +105,14 @@ export default {
     return {
       localItems: Array.isArray(this.items) ? [...this.items] : [],
       newItem: "",
-      draggingIndex: null,
-      itemRefs: {},
-      originalItems: null, // store original order in case drag gets cancelled
-      pointerId: null,
-      dragClone: null,
-      dragOffsetY: 0,
-      dragCorrectionX: 0,
-      dragCorrectionY: 0,
-      scrollParent: null,
-      autoScrollRAF: null,
-      autoScrollDelta: 0,
-      lastClientY: 0,
+      dragReorder: createDragReorder({
+        getOrder: () => [...this.localItems],
+        setOrder: (snapshot) => { this.localItems = snapshot; },
+      }),
     };
   },
   beforeUnmount() {
-    if (this.draggingIndex !== null || this.autoScrollRAF !== null) {
-      this.finishDrag(false);
-    }
+    this.dragReorder.cancel();
   },
   methods: {
     addItem() {
@@ -136,156 +126,6 @@ export default {
     },
     typeInfoFor(item) {
       return getTypeInfoFromExt(item);
-    },
-    onPointerDown(index, event) {
-      if (event.pointerType === "mouse" && event.button !== 0) return;
-      event.preventDefault();
-      this.draggingIndex = index;
-      this.originalItems = [...this.localItems];
-      this.pointerId = event.pointerId;
-      this.lastClientY = event.clientY;
-      // make all pointermove/up events for this pointer go to this element
-      event.target.setPointerCapture(event.pointerId);
-
-      const row = getObjectProperty(this.itemRefs, index);
-      if (row) {
-        const rect = row.getBoundingClientRect();
-        this.dragOffsetY = event.clientY - rect.top;
-        const clone = row.cloneNode(true);
-        clone.style.position = "fixed";
-        clone.style.left = `${rect.left}px`;
-        clone.style.top = `${rect.top}px`;
-        clone.style.width = `${rect.width}px`;
-        clone.style.margin = "0";
-        clone.style.pointerEvents = "none";
-        clone.style.zIndex = "9999";
-        clone.style.boxShadow = "0 4px 12px rgba(0, 0, 0, 0.25)";
-        row.parentNode.insertBefore(clone, row.nextSibling);
-        const actualRect = clone.getBoundingClientRect();
-        this.dragCorrectionX = rect.left - actualRect.left;
-        this.dragCorrectionY = rect.top - actualRect.top;
-        if (this.dragCorrectionX || this.dragCorrectionY) {
-          clone.style.left = `${rect.left + this.dragCorrectionX}px`;
-          clone.style.top = `${rect.top + this.dragCorrectionY}px`;
-        }
-        this.dragClone = clone;
-        this.scrollParent = this.getScrollParent(row);
-      }
-    },
-    onPointerMove(event) {
-      if (this.draggingIndex === null || event.pointerId !== this.pointerId) return;
-      event.preventDefault();
-      this.lastClientY = event.clientY;
-      if (this.dragClone) {
-        const top = event.clientY - this.dragOffsetY + (this.dragCorrectionY || 0);
-        this.dragClone.style.top = `${top}px`;
-      }
-      this.updateHoveredIndex(event.clientY);
-      this.updateAutoScroll(event.clientY);
-    },
-    onPointerUp(event) {
-      if (event.pointerId !== this.pointerId) return;
-      this.finishDrag(true);
-    },
-    onPointerCancel(event) {
-      if (event.pointerId !== this.pointerId) return;
-      this.finishDrag(false);
-    },
-    finishDrag(commit) {
-      // If drag was cancelled (not a clean pointerup), restore original order
-      if (!commit && this.originalItems !== null) {
-        this.localItems = this.originalItems;
-      }
-      this.originalItems = null;
-      this.draggingIndex = null;
-      this.pointerId = null;
-
-      if (this.dragClone) {
-        this.dragClone.remove();
-        this.dragClone = null;
-      }
-      if (this.autoScrollRAF) {
-        cancelAnimationFrame(this.autoScrollRAF);
-        this.autoScrollRAF = null;
-      }
-      this.autoScrollDelta = 0;
-      this.scrollParent = null;
-    },
-    updateHoveredIndex(clientY) {
-      if (this.draggingIndex === null) return;
-      const entries = Object.entries(this.itemRefs)
-        .map(([i, el]) => [Number(i), el])
-        .filter(([, el]) => !!el)
-        .sort((a, b) => a[0] - b[0]);
-
-      let targetIndex = entries.length;
-      for (const [i, el] of entries) {
-        if (i === this.draggingIndex) continue;
-        const rect = el.getBoundingClientRect();
-        if (clientY < rect.top + rect.height / 2) {
-          targetIndex = i;
-          break;
-        }
-      }
-      if (targetIndex !== this.draggingIndex) {
-        const array = [...this.localItems];
-        const [moved] = array.splice(this.draggingIndex, 1);
-        if (targetIndex > this.draggingIndex) targetIndex -= 1;
-        array.splice(targetIndex, 0, moved);
-        this.localItems = array;
-        this.draggingIndex = targetIndex;
-      }
-    },
-    getScrollParent(node) {
-      let el = node?.parentElement;
-      while (el) {
-        const style = getComputedStyle(el);
-        const scrollable =
-          (style.overflowY === "auto" || style.overflowY === "scroll") &&
-          el.scrollHeight > el.clientHeight;
-        if (scrollable) return el;
-        el = el.parentElement;
-      }
-      return document.scrollingElement || document.documentElement;
-    },
-    updateAutoScroll(clientY) {
-      const EDGE = 50;
-      const MAX_SPEED = 16;
-      const parent = this.scrollParent;
-      if (!parent) return;
-      const isWindowScroll = parent === document.scrollingElement || parent === document.documentElement;
-      const rect = isWindowScroll ? null : parent.getBoundingClientRect();
-      const top = isWindowScroll ? 0 : rect.top;
-      const bottom = isWindowScroll ? window.innerHeight : rect.bottom;
-
-      let delta = 0;
-      if (clientY < top + EDGE) {
-        delta = -MAX_SPEED * (1 - Math.max(0, clientY - top) / EDGE);
-      } else if (clientY > bottom - EDGE) {
-        delta = MAX_SPEED * (1 - Math.max(0, bottom - clientY) / EDGE);
-      }
-      this.autoScrollDelta = delta;
-      if (delta !== 0 && this.autoScrollRAF === null) {
-        this.autoScroll();
-      }
-    },
-    autoScroll() {
-      const isWindowScroll =
-        this.scrollParent === document.scrollingElement || this.scrollParent === document.documentElement;
-      const step = () => {
-        if (this.autoScrollDelta !== 0 && this.scrollParent) {
-          if (isWindowScroll) {
-            window.scrollBy(0, this.autoScrollDelta);
-          } else {
-            this.scrollParent.scrollTop += this.autoScrollDelta;
-          }
-          this.updateHoveredIndex(this.lastClientY);
-          this.autoScrollRAF = requestAnimationFrame(step);
-        } else {
-          this.autoScrollRAF = null;
-        }
-      };
-      this.autoScrollRAF = requestAnimationFrame(step);
     },
     cancel() {
       mutations.closeTopPrompt();
